+++ rbx_xml/src/types/mod.rs
@@ -29,6 +29,7 @@ mod rect;
 mod referent;
 mod security_capabilities;
 mod shared_string;
+mod smooth_grid;
 mod strings;
 mod tags;
 mod udims;
@@ -58,6 +59,7 @@ use self::{
     material_colors::write_material_colors,
     referent::{read_ref, write_ref},
     shared_string::{read_shared_string, write_shared_string},
+    smooth_grid::write_smooth_grid,
     tags::write_tags,
 };
 
@@ -118,6 +120,7 @@ macro_rules! declare_rbx_types {
                 Variant::Tags(value) => write_tags(writer, xml_property_name, value),
                 Variant::Attributes(value) => write_attributes(writer, xml_property_name, value),
                 Variant::MaterialColors(value) => write_material_colors(writer, xml_property_name, value),
+                Variant::SmoothGrid(value) => write_smooth_grid(writer, xml_property_name, value),
 
                 unknown => {
                     Err(writer.error(EncodeErrorKind::UnsupportedPropertyType(unknown.ty())))