+++ rbx_binary/src/tests/models.rs
@@ -69,5 +69,6 @@ binary_tests! {
     folder_with_enum_attribute,
     imagelabel_content,
     content_mixed,
+    physical_properties_acoustics,
     netassetref,
 }