+++ rbx_binary/CHANGELOG.md
@@ -1,6 +1,9 @@
 # rbx_binary Changelog
 
 ## Unreleased
+* Added support for `Terrain.SmoothGrid`, and adjacent support for `TerrainRegion.SmoothGrid`. ([#444])
+
+[#444]: https://github.com/rojo-rbx/rbx-dom/pull/444
 
 ## 1.0.0 (2025-03-28)
 * Dramatically improved performance of serializer and deserializer by using `Ustr` to represent property and class names ([#462]).