+++ rbx_binary/CHANGELOG.md
@@ -1,6 +1,7 @@
 # rbx_binary Changelog
 
 ## Unreleased
+* Implement support for `AcousticAbsorption` in `PhysicalProperties`.
 
 ## 1.0.0 (2025-03-28)
 * Dramatically improved performance of serializer and deserializer by using `Ustr` to represent property and class names ([#462]).