+++ rbx_binary/CHANGELOG.md
@@ -1,7 +1,9 @@
 # rbx_binary Changelog
 
 ## Unreleased
-* Implement support for `AcousticAbsorption` in `PhysicalProperties`.
+* Implement support for `AcousticAbsorption` in `PhysicalProperties`. ([#556])
+
+[#556]: https://github.com/rojo-rbx/rbx-dom/pull/556
 
 ## 1.0.0 (2025-03-28)
 * Dramatically improved performance of serializer and deserializer by using `Ustr` to represent property and class names ([#462]).