+++ rbx_binary/CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ## Unreleased
 * Implement support for serializing and deserializing the `NetAssetRef` type. ([#555])
+* Implement support for `AcousticAbsorption` in `PhysicalProperties`. ([#556])
 
 [#555]: https://github.com/rojo-rbx/rbx-dom/pull/555
+[#556]: https://github.com/rojo-rbx/rbx-dom/pull/556
 
 ## 1.0.0 (2025-03-28)
 * Dramatically improved performance of serializer and deserializer by using `Ustr` to represent property and class names ([#462]).